refact(resources): extract mobile header to its own component
