+++ src/components/MobileHeader/index.tsx
@@ -1,0 +1,18 @@
+import React from "react"
+import * as SC from "./styles"
+
+interface IMobileHeaderProps {
+  openMenu: () => void
+}
+
+export function MobileHeader({ openMenu }: IMobileHeaderProps) {
+  return (
+    <SC.MobileHeaderWrapper>
+      <SC.LogoWrapper>
+        <SC.StyledLogo />
+      </SC.LogoWrapper>
+      Resources
+      <SC.Burger onClick={openMenu} />
+    </SC.MobileHeaderWrapper>
+  )
+}
